+++ frontend/webapp/src/app/layout.tsx
@@ -2,7 +2,7 @@ import '@mantine/core/styles.css';
 import '@mantine/carousel/styles.css';
 import React from 'react';
 import { MantineProvider, ColorSchemeScript } from '@mantine/core';
-import PageLayout from '@/layout/PageLayout';
+import { PageLayout } from '@/layout';
 
 export const metadata = {
   title: 'Veekly Bobangs',
