+++ frontend/webapp/src/layout/index.tsx
@@ -1,0 +1,2 @@
+export { default as Navbar } from './Navbar';
+export { default as PageLayout } from './PageLayout';
